rename tests and fix import statements
